+++ CHANGELOG.md
@@ -5,6 +5,14 @@
 本项目遵循 [Keep a Changelog](https://keepachangelog.com/zh-CN/1.1.0/) 的格式，
 并遵循 [语义化版本号](https://semver.org/lang/zh-CN/)。
 
+## [1.5.0] - 2026-01-23
+
+### ✨ 新增
+- 支持对于 [PocketChat](https://github.com/PocketTogether/pocket-chat) 的转发 by @haruki1953 in https://github.com/Tweblog/tweblog/pull/5
+
+### 🔧 变更
+- 调整项目显示的文档地址，调整前端 forward-config 批量转发显示 by @haruki1953 in https://github.com/Tweblog/tweblog/pull/6
+
 ## [1.4.0] - 2026-01-22
 
 ### 🔧 变更
@@ -86,7 +94,7 @@
 - 新增 [批量导入](./Tweblog/src/guide/feature/tweet-import.md#批量导入)
 - 新增 [高级功能](./Tweblog/src/guide/feature/tweet-import.md#高级功能)：导入时可关联至转发记录
 - 导入、转发任务支持 [任务中止](./Tweblog/src/guide/feature/tweet-import.md#任务中止)
-- 新增 [自动转发](./Tweblog/src/guide/feature/tweet-forward.md#自动转发)
+- 新增 [批量转发](./Tweblog/src/guide/feature/tweet-forward.md#批量转发)
 - 新增 [转发记录设置](./Tweblog/src/guide/feature/tweet-forward.md#转发记录设置)
 
 ### 🧱 改进
@@ -105,6 +113,7 @@
 - 支持 X / Twitter 导入与转发
 - Web 版可作为个人博客使用
 
+[1.5.0]: https://github.com/Tweblog/tweblog/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/Tweblog/tweblog/compare/1.3.0...v1.4.0
 [1.3.0]: https://github.com/Tweblog/tweblog/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/Tweblog/tweblog/compare/1.1.0...1.2.0